Split generation use case emitters
